Add files via upload
ProjectGroup containing the VCL Windows UI app and a cmd line unit tests. I am satisifed that this shows my deve,lopment style, but obviously there are many directions in which this could be improved, most notably providing more test data and writing more Unit tests. Another area that I did not address as there was no need, is the optimisation. Currently all the queries simply iterate through all the data to get results, which is fine for the limited usage of our case, but when dealing with millions of services, could prove a problem. It is fairly easy to code indexing inside TBusServiceData class and I can explain how I'd do it in the upcoming interview ..